requirements.txt
apache-airflow[mysql,celery,rabbitmq]==1.10.12
cattrs==1.0.0